requirements.txt
annotated-types==0.5.0
black==23.7.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pydantic==2.0.3
pydantic_core==2.3.0
python-dotenv==1.0.0
rel==0.4.9
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.3
websocket-client==1.6.1
